pyproject.toml
[project]
name = "backfill"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "", email = "febriyantabidin@gmail.com"},
]
dependencies = [
    "pandas>=2.2.3",
    "sqlalchemy>=2.0.38",
    "mysql-connector-python>=9.2.0",
    "dotenv>=0.9.9",
]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
